+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+### Fixed
+
+ - [ET-2131] Fix key packets missing error when sending messages.
+
+
 ## [0.55.3] - 2025-03-13
 
 ### Fixed